requirements.txt
hidpi_tk==1.3.1
pywebview==6.1
